+++ fili-core/src/test/groovy/com/yahoo/bard/webservice/web/apirequest/generator/metric/ProtocolLogicalMetricGeneratorSpec.groovy
@@ -22,6 +22,8 @@ import com.yahoo.bard.webservice.data.metric.protocol.ProtocolMetricImpl
 import com.yahoo.bard.webservice.data.metric.protocol.ProtocolSupport
 import com.yahoo.bard.webservice.data.time.DefaultTimeGrain
 import com.yahoo.bard.webservice.data.time.Granularity
+import com.yahoo.bard.webservice.druid.model.aggregation.Aggregation
+import com.yahoo.bard.webservice.druid.model.aggregation.LongSumAggregation
 import com.yahoo.bard.webservice.table.LogicalTable
 import com.yahoo.bard.webservice.table.PhysicalTable
 import com.yahoo.bard.webservice.table.TableGroup
@@ -41,24 +43,29 @@ class ProtocolLogicalMetricGeneratorSpec extends Specification {
     Protocol testProtocol
     ProtocolDictionary protocolDictionary
 
-    TemplateDruidQuery emptyTdq
+    TemplateDruidQuery testTdq
 
     String baseMetadataTransformMetricName
     String rawMetadataTransformMetricName
+    Aggregation baseMetadataAggregation
     ProtocolMetric baseMetric
 
     String validLogicalMetricName
+    Aggregation validLogicalMetricAggregation
     LogicalMetric validLogicalMetric
 
     String invalidLogicalMetricName
+    Aggregation invalidLogicalMetricAggregation
     LogicalMetric invalidLogicalMetric
 
     String validProtocolMetricBaseName
+    Aggregation validProtocolMetricAggregation
     String validProtocolMetricRawName
     ProtocolMetric validProtocolMetricPreTransform
     ProtocolMetric validProtocolMetricPostTransform
 
     String invalidProtocolMetricBaseName
+    Aggregation invalidProtocolMetricAggregation
     String invalidProtocolMetricRawName
     ProtocolMetric invalidProtocolMetricPreTransform
     ProtocolMetric invalidProtocolMetricPostTransform
@@ -91,62 +98,78 @@ class ProtocolLogicalMetricGeneratorSpec extends Specification {
         // setup test metrics
         baseMetadataTransformMetricName = "baseMetric"
         rawMetadataTransformMetricName = "baseMetric(p=unused)"
-        emptyTdq = new TemplateDruidQuery([], [])
+        baseMetadataAggregation = new LongSumAggregation(baseMetadataTransformMetricName, "base_metric_field")
+
+        validLogicalMetricName = "validLogicalMetric"
+        validLogicalMetricAggregation = new LongSumAggregation(validLogicalMetricName, "valid_logical_metric_field")
+
+        invalidLogicalMetricName = "invalidLogicalMetric"
+        invalidLogicalMetricAggregation = new LongSumAggregation(invalidLogicalMetricName, "invalid_logical_metric_field")
+
+        validProtocolMetricBaseName = "validProtocolMetric"
+        validProtocolMetricAggregation = new LongSumAggregation(validProtocolMetricBaseName, "valid_protocol_metric_field")
+        validProtocolMetricRawName = "RAW_validProtocolMetric"
+
+        invalidProtocolMetricBaseName = "invalidProtocolMetric"
+        invalidProtocolMetricAggregation = new LongSumAggregation(invalidProtocolMetricBaseName, "invalid_protocol_metric_field")
+        invalidProtocolMetricRawName = "RAW_invalidProtocolMetric"
+
+        testTdq = new TemplateDruidQuery(
+                [
+                        baseMetadataAggregation,
+                        validLogicalMetricAggregation,
+                        invalidLogicalMetricAggregation,
+                        validProtocolMetricAggregation,
+                        invalidProtocolMetricAggregation
+                ],
+                []
+        )
         baseMetric = new ProtocolMetricImpl(
                 new LogicalMetricInfo(baseMetadataTransformMetricName),
-                emptyTdq,
+                testTdq,
                 TEST_RSM,
                 new ProtocolSupport([testProtocol]),
         )
         metricDictionary = new MetricDictionary()
         metricDictionary.add(baseMetric)
 
-        // Prepare test metrics
-        validLogicalMetricName = "validLogicalMetric"
         validLogicalMetric = new LogicalMetricImpl(
                 new LogicalMetricInfo(validLogicalMetricName),
-                emptyTdq,
+                testTdq,
                 TEST_RSM
         )
         metricDictionary.add(validLogicalMetric)
 
-        invalidLogicalMetricName = "invalidLogicalMetric"
         invalidLogicalMetric = new LogicalMetricImpl(
                 new LogicalMetricInfo(invalidLogicalMetricName),
-                emptyTdq,
+                testTdq,
                 TEST_RSM
         )
         metricDictionary.add(invalidLogicalMetric)
 
-        validProtocolMetricBaseName = "validProtocolMetric"
-        validProtocolMetricRawName = "RAW_validProtocolMetric"
-
         validProtocolMetricPreTransform = new ProtocolMetricImpl(
                 new LogicalMetricInfo(validProtocolMetricBaseName),
-                emptyTdq,
+                testTdq,
                 TEST_RSM
         )
         metricDictionary.add(validProtocolMetricPreTransform)
 
         validProtocolMetricPostTransform = new ProtocolMetricImpl(
                 new GeneratedMetricInfo(validProtocolMetricRawName, validProtocolMetricBaseName),
-                emptyTdq,
+                testTdq.renameMetricField(validProtocolMetricBaseName, validProtocolMetricRawName),
                 TEST_RSM
         )
 
-        invalidProtocolMetricBaseName = "invalidProtocolMetric"
-        invalidProtocolMetricRawName = "RAW_invalidProtocolMetric"
-
         invalidProtocolMetricPreTransform = new ProtocolMetricImpl(
                 new LogicalMetricInfo(invalidProtocolMetricBaseName),
-                emptyTdq,
+                testTdq,
                 TEST_RSM
         )
         metricDictionary.add(invalidProtocolMetricPreTransform)
 
         invalidProtocolMetricPostTransform = new ProtocolMetricImpl(
                 new GeneratedMetricInfo(invalidProtocolMetricRawName, invalidProtocolMetricBaseName),
-                emptyTdq,
+                testTdq.renameMetricField(invalidProtocolMetricBaseName, invalidProtocolMetricRawName),
                 TEST_RSM
         )